First successful read of descriptors from vmulti.hid file.
This is still suboptimal. For example, getting report descriptor involves reading vmulti.hid twice. Code is also very messy. However, getting it to work is a good sign.

An example vmulti.hid has also been committed, in order to facilitate easier testing on an actual and functional descriptor, which has an added boon of not including extra devices, easing uninstallation on Windows XP.